+++ studio-idea-plugin/src/main/java/com/evolveum/midpoint/studio/action/TestConnectionAction.java
@@ -6,12 +6,11 @@ import com.evolveum.midpoint.studio.impl.EnvironmentService;
 import com.evolveum.midpoint.studio.impl.MidPointClient;
 import com.evolveum.midpoint.studio.impl.ShowExceptionNotificationAction;
 import com.evolveum.midpoint.studio.impl.cache.EnvironmentCacheManager;
+import com.evolveum.midpoint.studio.ui.StudioAction;
 import com.evolveum.midpoint.studio.util.MidPointUtils;
 import com.evolveum.midpoint.studio.util.RunnableUtils;
 import com.intellij.notification.NotificationAction;
 import com.intellij.notification.NotificationType;
-import com.intellij.openapi.actionSystem.ActionUpdateThread;
-import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.progress.ProgressIndicator;
 import com.intellij.openapi.progress.ProgressManager;
@@ -20,40 +19,21 @@ import com.intellij.openapi.progress.impl.BackgroundableProcessIndicator;
 import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.NotNull;
 
-import javax.swing.*;
-
 /**
  * Created by Viliam Repan (lazyman).
  */
-public class TestConnectionAction extends AnAction {
+public class TestConnectionAction extends StudioAction {
 
     private static final String NOTIFICATION_KEY = "Test Connection";
 
     @Override
-    public @NotNull ActionUpdateThread getActionUpdateThread() {
-        return ActionUpdateThread.BGT;
-    }
-
-    @Override
-    public void update(@NotNull AnActionEvent e) {
-        super.update(e);
-
-        Project project = e.getProject();
-        if (project == null) {
-            SwingUtilities.invokeLater(() -> e.getPresentation().setEnabled(false));
-            return;
+    protected boolean isEnabled(@NotNull AnActionEvent e) {
+        if (!super.isEnabled(e)) {
+            return false;
         }
 
-        boolean hasFacet = MidPointUtils.hasMidPointFacet(e.getProject());
-        if (!hasFacet) {
-            SwingUtilities.invokeLater(() -> e.getPresentation().setEnabled(false));
-            return;
-        }
-
-        EnvironmentService em = EnvironmentService.getInstance(project);
-        Environment selected = em.getSelected();
-
-        SwingUtilities.invokeLater(() -> e.getPresentation().setEnabled(selected != null));
+        EnvironmentService em = EnvironmentService.getInstance(e.getProject());
+        return em.getSelected() != null;
     }
 
     @Override
@@ -62,7 +42,8 @@ public class TestConnectionAction extends AnAction {
         EnvironmentService em = EnvironmentService.getInstance(project);
         Environment selected = em.getSelected();
 
-        Task.Backgroundable task = new Task.Backgroundable(e.getProject(), "Testing connection for '" + selected.getName() + "'") {
+        Task.Backgroundable task = new Task.Backgroundable(
+                e.getProject(), "Testing connection for '" + selected.getName() + "'") {
 
             @Override
             public void run(@NotNull ProgressIndicator indicator) {
@@ -75,7 +56,8 @@ public class TestConnectionAction extends AnAction {
                 }.run();
             }
         };
-        ProgressManager.getInstance().runProcessWithProgressAsynchronously(task, new BackgroundableProcessIndicator(task));
+        ProgressManager.getInstance()
+                .runProcessWithProgressAsynchronously(task, new BackgroundableProcessIndicator(task));
     }
 
     private void testConnection(Project project, Environment environment) {