+++ studio-idea-plugin/src/main/java/com/evolveum/midpoint/studio/impl/configuration/CleanupConfiguration.java
@@ -6,6 +6,7 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 public class CleanupConfiguration implements Serializable {
 
@@ -80,6 +81,7 @@ public class CleanupConfiguration implements Serializable {
         this.removeContainerIds = removeContainerIds;
     }
 
+    @NotNull
     public MissingRefObjects getMissingReferences() {
         if (missingReferences == null) {
             missingReferences = new MissingRefObjects();
@@ -127,7 +129,7 @@ public class CleanupConfiguration implements Serializable {
 
         List<CleanupPathConfiguration> paths = getCleanupPaths().stream()
                 .map(CleanupPathConfiguration::copy)
-                .toList();
+                .collect(Collectors.toList());  // can't use .toList() directly because it would be immutable list
         configuration.setCleanupPaths(paths);
 
         configuration.setCleanupConnectorReferences(cleanupConnectorReferences);